Fix for 1-second-bug
Not sure if it’s really more of a workaround, but it works.
When the timer is idle and you tap it to switch between remaining and
not remaining mode, a second used to appear out of nowhere. With this
fix, the second no longer appears, but the timer stays at 0:00.